+++ Tests/Ifrit_Levenstein/Levenstein_TestsCore.swift
@@ -1,3 +1,4 @@
+
 import Foundation
 import XCTest
 @testable import Ifrit
@@ -18,25 +19,25 @@ final class Levenstein_TestsCore: XCTestCase {
         )
         XCTAssertEqual(actual, expected)
     }
-
+    
     func testContiguousCharactersHaveHigherScore() {
         let a1 = bestMatch(query: "pp", input: "pickled pepper")!
         let a2 = bestMatch(query: "pp", input: "Pied Piper")!
         XCTAssertTrue(a1.score > a2.score)
     }
-
+    
     func testCharactersAtBeginningOfWordsHaveHigherScore() {
         let a1 = bestMatch(query: "pp", input: "Pied Piper")!
         let a2 = bestMatch(query: "pp", input: "porcupine")!
         XCTAssertTrue(a1.score > a2.score)
     }
-
+    
     func testCamelCaseHumpsHaveHigherScore() {
         let a1 = bestMatch(query: "bm", input: "BatMan")!
         let a2 = bestMatch(query: "bm", input: "Batman")!
         XCTAssertTrue(a1.score > a2.score)
     }
-
+    
     func testFirstLettersOfWordsHaveHigherScore() {
         let a1 = bestMatch(query: "bm", input: "Bat man")!
         let a2 = bestMatch(query: "bm", input: "Batman")!